database/requirements.txt
Flask==2.1.0
Werkzeug==2.2.2
google-cloud-firestore==2.4.0
gunicorn==20.1.0
